+++ components/core/kmpparcelize/src/androidMain/kotlin/com/flipperdevices/core/kmpparcelize/KMPParcelize.kt
@@ -1,0 +1,5 @@
+package com.flipperdevices.core.kmpparcelize
+
+import kotlinx.android.parcel.Parcelize
+
+actual typealias KMPParcelize = Parcelize