+++ components/keyedit/api/src/main/java/com/flipperdevices/keyedit/api/NotSavedFlipperKey.kt
@@ -1,30 +1,26 @@
 package com.flipperdevices.keyedit.api
 
 import android.content.Context
-import android.os.Parcelable
 import com.flipperdevices.bridge.dao.api.model.FlipperFile
 import com.flipperdevices.bridge.dao.api.model.FlipperFilePath
 import com.flipperdevices.bridge.dao.api.model.FlipperKeyContent
 import com.flipperdevices.core.ktx.jre.FlipperDispatchers
 import com.flipperdevices.core.preference.FlipperStorageProvider
 import kotlinx.coroutines.withContext
-import kotlinx.parcelize.Parcelize
 import kotlinx.serialization.Serializable
 
-@Parcelize
 @Serializable
 data class NotSavedFlipperKey(
     val mainFile: NotSavedFlipperFile,
     val additionalFiles: List<NotSavedFlipperFile>,
     val notes: String?
-) : Parcelable
+)
 
-@Parcelize
 @Serializable
 data class NotSavedFlipperFile(
     val path: FlipperFilePath,
     val content: FlipperKeyContent
-) : Parcelable
+)
 
 suspend fun FlipperFile.toNotSavedFlipperFile(
     context: Context